refactor: User 엔티티, 디렉토리 이동
