+++ src/main/java/roomescape/user/domain/dto/UserRequestDto.java
@@ -1,5 +1,7 @@
 package roomescape.user.domain.dto;
 
+import roomescape.user.domain.User;
+
 public record UserRequestDto(String name, String email, String password) {
 
     public User toEntity() {
