+++ src/main/java/roomescape/user/domain/dto/UserResponseDto.java
@@ -1,5 +1,7 @@
 package roomescape.user.domain.dto;
 
+import roomescape.user.domain.User;
+
 public record UserResponseDto(Long id, String name, String email, String password) {
 
     public static UserResponseDto of(User user) {
